Fix new npgsql enum conversion problem
Now enums are explicitly converted into numbers on insert into DB.
Also updated sequences import repositories.

+++ LibiadaWeb/Models/Repositories/Sequences/FmotifRepository.cs
@@ -88,7 +88,7 @@
                     }
                 }
             }
-            
+
             return Create(fmotif, notes, fmotifChain.Building);
         }
 
@@ -167,7 +167,7 @@
                 {
                     ParameterName = "notation",
                     NpgsqlDbType = NpgsqlDbType.Smallint,
-                    Value = Notation.FormalMotifs
+                    Value = (byte)Notation.FormalMotifs
                 },
                 new NpgsqlParameter
                 {
@@ -185,7 +185,7 @@
                 {
                     ParameterName = "fmotif_type",
                     NpgsqlDbType = NpgsqlDbType.Smallint,
-                    Value = fmotif.Type
+                    Value = (byte)fmotif.Type
                 }
             };
             return parameters;
